Add Arangodb
This is a bit of a busy commit so it's worth explaining some things.
The thrust of the issue is talking to Arango which requires some env vars.
Those env vars are managed & verified by dotenv-safe based on the values in
.env.example.

Server, db and initialization concerns have been teased apart to make testing
easy, and finally since Jest runs your tests in parallel there is some code in
src/utils.js to create a database per test file so that tests don't interfere
with each other by deleting each others data.

The end result there is simply that we have the same "hello world" message, but
this time coming from the database.