Add Sealed Secrets (#97)
This commit adds Sealed Secrets to allow us to store our secrets alongside the
rest of our project on Github.

The idea is a public/private key pair is created, where the private key lives
in the cluster. The public key is stored in the root directory under the name
`publickey.crt`. Secrets can then be created and stored with everything else on
Github simplifying deployments.

Creating a new secret with the key requires the kubeseal command to be installed.
A new secret could then be created with a command like this:

```bash
kubectl create secret generic foo --from-literal=foo=bar \
 --dry-run -o yaml | kubeseal --cert publickey.crt --format yaml
apiVersion: bitnami.com/v1alpha1
kind: SealedSecret
metadata:
  creationTimestamp: null
  name: foo
  namespace: default
spec:
  encryptedData:
    foo: AgClberDBr0YfUjjt0px37S3BS57Fbv2YCSg6....
```

The value of foo is first encrypted with the public key, and then base64 encoded. With that in mind to verify the value is encrypted requires you to reverse the base64 encoding. I found the yq tool (pip install yq) to be helpful (except it adds quotes I need to remove with tr), but to give a sense of it:

```bash
kubectl create secret generic foo --from-literal=foo=bar \
 --dry-run -o yaml | kubeseal --cert publickey.crt --format yaml | \
 yq '.spec.encryptedData.foo' - | tr -d \" | base64 -d -
m��~[���l!?A�|�lТ�t�zf��XԘ����qP�
```