The denouement page! (#217)
This is currently a page at /old/thanks that displays fixed data from the api
This commit also pins the version of react-apollo to 2.5.2 as 2.5.3 throws an
error:

`Could not find "client" in the context of ApolloConsumer. Wrap the root
component in an <ApolloProvider>`

Similarly, clean-tag 3.1.1 appears to be broken so pinning to the 3.0.0 version to resolve the

`TypeError: Cannot convert undefined or null to object at Function.keys` error.

Pinning the versions like this is temporary and will be revisited shortly.