Use better summary data
This commit modifies the frontend to make use of the new flaggingsWithin
function exposed by the api. This will address some of the shortcomings with
the data visualization (ie, not displaying days with 0 reports).
The values for the demo are still hardcoded (ie the phone number identifier),
but the refactoring here will allow us to remove them and get this working for
real easily. This will be the focus of a followup PR.