Serve better summary data (#267)
This commit adds a new function to the graphql api that returns proper data for
the summary graph. "Proper" in this case meaning that it the user supplies a
date range and the summary includes days with 0 reports within the range.

This addresses #255 and will be followed with a commit to get the frontend
using this new function.