Clean up cache creation (#262)
* Clean up cache creation

The Apollo cache was being created with default values for the narrative
gathering tool. This commit removes those and defaults the cached value for
language to "fr" if it can find it in the navigator.languages array.

* Fix language check

The order of `navigator.languages` reflects the users language preferences,
with `navigator.language` representing the users first choice. This check
starts with language, and then falls back to searching through languages.