Refactor cloudbuild file
This commit reworks the cloudbuild file, moving test credentials and such into
variables and adding a few missing variables.

You may also note that we are now using Kaniko for image building. This is a
new tools that lets you build containers without docker, and it should give a
bit of a speed boost as well as making our images reproducable (ie, two images
with the same content build at different times would have the same SHA).

Also worth mentioning is the removal of the images field, which isn't needed
with
[kaniko](https://github.com/GoogleCloudPlatform/cloud-builders-community/issues/212#issuecomment-475211337).