Move Kubernetes config into a separate folder (#327)
* Move Kubernetes config into a separate folder

This commit adds the platform folder which now holds all config. This config is
set up to use kustomize so that we can generate variations on it for various
environments we use.

* Remove istio injection label.

* Remove compliance checks... for now.

The current compliance setup assumes istio is installed, which is no longer
true. Removing for now and will revisit shortly.