Update image tags in the base not overlays
Flux now allows manifests to be generated, and file that controls that is .flux.yaml:
https://github.com/weaveworks/flux/blob/master/site/fluxyaml-config-files.md#abstracting-out-file-access-generators-and-updaters

We are using Kustomize to generate and update our manifests, but were only
updating the images for the AKS overlay, which made it very confusing when
trying things in Minikube.

This commit changes the updates so they are applied to the kustomization.yaml
file in the base folder, which means they will be available to all the other
overlays.