Fix minio
This commit adds the missing config to finally get minio working.
Problem one was the fact that the service did not use the right selector, and
therefore created a dns entry within k8s that applied to 0 pods.
This resulted in not being able to connect to minio even though it was running
properly and the dns would resolve.

The second issue was the missing config. You'll notice that values from the
minio-secrets config is just being reused for the api.

Lastly, the bucket name was wrong.

Anything you upload to the api should be visible with the following command:
```sh
gsutil ls gs://report-a-cybercrime-alpha-uploads
```

You can test the uploads with curl like this:

```sh
$ curl "http://192.168.39.125:31656/graphql" \
 -F operations='{ "query": "mutation ($file: Upload!) { uploadFile(file: $file) {success} }", "variables": { "file": null } }' -F map='{ "0": ["variables.file"] }' \
 -F 0=@api/src/__tests__/kitten.jpg
```