Replace getByText and getByLabelText (#755)
These functions made it difficult to evolve the content without test breakage.
This commit removes almost all usage of these functions. The remaining cases
are either in p1 (soon to be deleted), the components folder (not effected by
content changes), or are cases where the text being searched for is supplied in
the actual test.